scope exports of icon paths
