+++ packages/perseus-editor/src/components/viewport-resizer.tsx
@@ -2,14 +2,15 @@
  * A component that displays controls for choosing a viewport size.
  * Renders three buttons: "Phone", "Tablet", and "Desktop".
  */
-import {components, constants, icons} from "@khanacademy/perseus";
+import {components, constants} from "@khanacademy/perseus";
 import * as React from "react";
+
+import {iconDesktop, iconMobilePhone, iconTablet} from "../styles/icon-paths";
 
 import type {DeviceType} from "@khanacademy/perseus";
 
 const {ButtonGroup, InlineIcon} = components;
 const {devices} = constants;
-const {iconDesktop, iconMobilePhone, iconTablet} = icons;
 
 type Props = {
     /** The current device type that is selected. */
