+++ packages/perseus/src/index.ts
@@ -47,7 +47,7 @@ export * as constants from "./components/constants";
 export * as globalStyles from "./styles/global-styles";
 // TODO(FEI-4504): Replace globalStyles with wonder-blocks where possible
 export * as globalConstants from "./styles/global-constants";
-export * as icons from "./icon-paths";
+export {iconChevronDown, iconTrash} from "./icon-paths";
 export * as Dependencies from "./dependencies";
 export {Log} from "./logging/log";
 export {default as JiptParagraphs} from "./jipt-paragraphs";
